Solving policyApiUrl prop and pom dependencies version
