+++ pom.xml
@@ -9,7 +9,7 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>jmgomez</groupId>
-	<artifactId>ApiPolicy</artifactId>
+	<artifactId>PolicyApi</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>ApiPolicy</name>
 	<description>Demo project for Spring Boot</description>
@@ -31,6 +31,8 @@
 		<spring-cloud.version>Hoxton.SR9</spring-cloud.version>
 		<org.mapstruct.version>1.6.0</org.mapstruct.version>
 		<lombok.version>1.18.34</lombok.version>
+		<lombok.mapstruct.version>0.2.0</lombok.mapstruct.version>
+		<cache.version>3.1.5</cache.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -62,7 +64,7 @@
 		<dependency>
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok-mapstruct-binding</artifactId>
-			<version>0.2.0</version>
+			<version>${lombok.mapstruct.version}</version>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.security</groupId>
@@ -85,7 +87,7 @@
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-cache</artifactId>
-			<version>3.1.5</version>
+			<version>${cache.version}</version>
 		</dependency>
 	</dependencies>
 	<dependencyManagement>
@@ -115,7 +117,7 @@
 						<path>
 							<groupId>org.projectlombok</groupId>
 							<artifactId>lombok-mapstruct-binding</artifactId>
-							<version>0.2.0</version>
+							<version>${lombok.mapstruct.version}</version>
 						</path>
 						<path>
 							<groupId>org.projectlombok</groupId>
