+++ src/main/java/jmgomez/policyapi/config/RequestInterceptor.java
@@ -1,9 +1,9 @@
-package jmgomez.apipolicy.config;
+package jmgomez.policyapi.config;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import jmgomez.apipolicy.service.PolicyService;
+import jmgomez.policyapi.service.PolicyService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.security.core.context.SecurityContextHolder;
