+++ src/main/java/jmgomez/policyapi/model/Accident.java
@@ -1,4 +1,4 @@
-package jmgomez.apipolicy.model.dto;
+package jmgomez.policyapi.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class AccidentDto {
-    @JsonProperty("sinisterId")
+public class Accident {
+    @JsonProperty("siniestroId")
     private String sinisterId;
-    @JsonProperty("status")
+    @JsonProperty("estado")
     private String status;
 }
