+++ src/main/java/jmgomez/policyapi/model/dto/AccidentDto.java
@@ -1,4 +1,4 @@
-package jmgomez.apipolicy.model;
+package jmgomez.policyapi.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class Policy {
-    @JsonProperty("polizaId")
-    private String policyId;
-    @JsonProperty("descripcion")
-    private String description;
+public class AccidentDto {
+    @JsonProperty("sinisterId")
+    private String sinisterId;
+    @JsonProperty("status")
+    private String status;
 }
