+++ src/main/java/jmgomez/policyapi/model/dto/PolicyDto.java
@@ -1,4 +1,4 @@
-package jmgomez.apipolicy.model;
+package jmgomez.policyapi.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class Accident {
-    @JsonProperty("siniestroId")
-    private String sinisterId;
-    @JsonProperty("estado")
-    private String status;
+public class PolicyDto {
+    @JsonProperty("policyId")
+    private String policyId;
+    @JsonProperty("description")
+    private String description;
 }
