+++ src/main/resources/application.properties
@@ -1,2 +1,2 @@
 spring.application.name=apiPolicy
-spring.application.policyServiceUrl=http://localhost:8081
+policyApi.services.policies.baseURL=http://localhost:8081
